pyproject.toml
[project]
name = "rootagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fakeredis==2.32.1",
    "fastapi==0.124.2",
    "httpx==0.28.1",
    "jinja2==3.1.6",
    "litellm==1.80.9",
    "passlib[bcrypt]==1.7.4",
    "pydantic==2.12.5",
    "pytest==9.0.2",
    "python-dotenv==1.2.1",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.20",
    "redis==7.1.0",
    "smolagents==1.23.0",
    "uvicorn==0.38.0",
]
